Update firmware check and handling in update script for kernel 6.6.72-v8+

+++ update.sh
@@ -61,12 +61,13 @@ run_techbase_update() {
     fi
   else
     printf "\n${red}Klarte ikke å utføre kommandoen: $1${clear}!"
+    return -1
   fi
 }
 
 if [ "$(uname -r)" = "6.6.72-v8+" ]; then
     echo "Running on 6.6.72-v8+ kernel"
-    run_techbase_update "timeout 120 softmgr update firmware -b x500_6.6.72-beta"
+    run_techbase_update "timeout 120 softmgr check firmware -b x500_6.6.72-beta"
     if [ $? -eq 0 ]; then
         echo "Firmware up to date"
         
@@ -77,11 +78,24 @@ if [ "$(uname -r)" = "6.6.72-v8+" ]; then
         run_techbase_update "timeout 120 softmgr update core -b x500_6.6.72-beta"
         run_techbase_update "timeout 120 softmgr update all -b x500_6.6.72-beta"
     else
+        run_techbase_update "timeout 120 softmgr update firmware -b x500_6.6.72-beta"
+        if [ $? -eq 1 ]; then
         echo "Firmware updated successfully - Will reboot now"
         green='\033[0;32m'
         clear='\033[0m'
         printf "\n${green}Kjør update på nytt etter omstart${clear}!"
-        reboot        
+        reboot    
+        fi
+    fi
+else
+# Run the firmware update command with a timeout
+run_techbase_update "timeout 120 softmgr update firmware -f yes"
+run_techbase_update "timeout 120 softmgr update core -f yes"
+run_techbase_update "timeout 120 softmgr update lib -f yes"
+run_techbase_update "timeout 30 softmgr update all"
+fi
+
+cp dhcpcd.backup /etc/dhcpcd.conf
     fi
 else
 # Run the firmware update command with a timeout
